requirements.txt
pandas==1.2.4
numpy==1.19.5
nibabel==3.2.1
colorama==0.4.3
scipy==1.7.0
pingouin==0.5.0
statsmodels==0.13.1
python-decouple==3.6
